client: add explicit UserAgent, remove Options and constructor
The Options and New serve no purpose it's just copying subset of one struct to another, and the target struct have all the fields exported anyway.

Header.Set converts the header name to the canonical form so there is no point in making them lowercase (which is not correct btw).

Default user agent in a library is a bad idea.
Library is not a user callers are.